reset bounds after updating size
I want to build a docking component for floating windows that works like windows. I want to be able to dock a component and then, if I pull the component away, I want it to shink. If the bounds aren't updated to take the shrinking of the component into account, it'll cause some strange boundary issues.